Extract BuildingExtrusionStyler to be part of NavigationMapView. Remove unused code.
